refactor: 일정·할 일 공통 UI 폴더 위치 정리
