+++ src/shared/ui/Modals/ScheduleEditor/ScheduleDateTimeSection.tsx
@@ -6,9 +6,9 @@ import { useScheduleCalendarOverlay } from '@/shared/hooks/addSchedule'
 import { useCalendarFieldPicker } from '@/shared/hooks/form/useCalendarFieldPicker'
 import type { ScheduleEditorFormValues } from '@/shared/types/event/event'
 import type { ScheduleEditorFormProps } from '@/shared/types/modal/scheduleEditor'
+import CustomDatePicker from '@/shared/ui/calendar/CustomDatePicker/CustomDatePicker'
+import CustomTimePicker from '@/shared/ui/calendar/CustomTimePicker/CustomTimePicker'
 import Checkbox from '@/shared/ui/common/Checkbox/Checkbox'
-import CustomDatePicker from '@/shared/ui/common/CustomDatePicker/CustomDatePicker'
-import CustomTimePicker from '@/shared/ui/common/CustomTimePicker/CustomTimePicker'
 import * as S from '@/shared/ui/Modals/ScheduleEditor/index.style'
 import { formatDisplayDate } from '@/shared/utils/date'
