+++ src/shared/ui/Modals/TodoEditor/TodoDateTimeSection.tsx
@@ -5,9 +5,9 @@ import { useFormContext, useWatch } from 'react-hook-form'
 import { useTodoCalendarOverlay } from '@/shared/hooks/addTodo'
 import type { TodoEditorFormValues } from '@/shared/types/event/event'
 import type { TodoEditorFormProps } from '@/shared/types/modal/todoEditor'
+import CustomDatePicker from '@/shared/ui/calendar/CustomDatePicker/CustomDatePicker'
+import CustomTimePicker from '@/shared/ui/calendar/CustomTimePicker/CustomTimePicker'
 import Checkbox from '@/shared/ui/common/Checkbox/Checkbox'
-import CustomDatePicker from '@/shared/ui/common/CustomDatePicker/CustomDatePicker'
-import CustomTimePicker from '@/shared/ui/common/CustomTimePicker/CustomTimePicker'
 import * as S from '@/shared/ui/Modals/TodoEditor/index.style'
 import { formatDisplayDate } from '@/shared/utils/date'
