+++ src/shared/ui/calendar/CustomDatePicker/CustomDatePicker.style.ts
