+++ src/shared/ui/scheduleTodo/RecentSearch/RecentSearch.style.ts
