+++ src/shared/ui/scheduleTodo/RepeatTypeGroup/RepeatTypeGroup.style.ts
