+++ src/shared/ui/scheduleTodo/SearchPlace/SearchPlace.style.ts
