+++ src/shared/ui/scheduleTodo/SelectColor/SelectColor.style.ts
