+++ src/shared/ui/scheduleTodo/TerminationPanel/TerminationPanel.style.ts
