+++ src/shared/ui/scheduleTodo/TerminationPanel/TerminationPanel.tsx
@@ -11,8 +11,8 @@ import { createPortal } from 'react-dom'
 import { theme } from '@/shared/styles/theme'
 import type { RepeatConfigSchema } from '@/shared/types/event/event'
 import type { RepeatConfig } from '@/shared/types/recurrence/repeat'
+import CustomDatePicker from '@/shared/ui/calendar/CustomDatePicker/CustomDatePicker'
 import Checkbox from '@/shared/ui/common/Checkbox/Checkbox'
-import CustomDatePicker from '@/shared/ui/common/CustomDatePicker/CustomDatePicker'
 import { formatIsoDate } from '@/shared/utils/date'
 
 import * as S from './TerminationPanel.style'
