+++ src/shared/ui/scheduleTodo/TitleSuggestionInput/TitleSuggestionInput.style.ts
